modified ClassicalSimulationExecutor's qapply result to dict.
